+++ src/pages/Login.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import { doLogin } from '../actions/index';
-import { getToken } from '../services/api';
+import { Link, Redirect } from 'react-router-dom';
+import { doLogin, getQuestionsToStore } from '../actions/index';
+import { getToken, getQuestions } from '../services/api';
 import './Login.css';
 
 class Login extends React.Component {
@@ -12,6 +12,7 @@ class Login extends React.Component {
     this.state = {
       invalidName: true,
       invalidEmail: true,
+      loginReady: false,
       name: '',
       email: '',
     };
@@ -19,12 +20,17 @@ class Login extends React.Component {
     this.changeHandler = this.changeHandler.bind(this);
     this.validateBtnLogin = this.validateBtnLogin.bind(this);
     this.enableButton = this.enableButton.bind(this);
-    this.getAndSaveToken = this.getAndSaveToken.bind(this);
+    this.getAndSaveQuestions = this.getAndSaveQuestions.bind(this);
   }
 
-  async getAndSaveToken() {
+  async getAndSaveQuestions() {
     const token = await getToken();
     localStorage.setItem('token', token);
+    const { saveQuestions } = this.props;
+    const API_RESULT = await getQuestions(token);
+    console.log(API_RESULT);
+    saveQuestions(API_RESULT);
+    this.setState({ loginReady: true });
   }
 
   validateBtnLogin() {
@@ -55,8 +61,11 @@ class Login extends React.Component {
   }
 
   render() {
-    const { name, email } = this.state;
+    const { name, email, loginReady } = this.state;
     const { doFormLogin } = this.props;
+    if (loginReady) {
+      return <Redirect to="/trivia" />;
+    }
     return (
       <main>
         <input
@@ -77,19 +86,17 @@ class Login extends React.Component {
           name="email"
           onChange={ (e) => { this.changeHandler(e); } }
         />
-        <Link to="/trivia">
-          <button
-            type="button"
-            data-testid="btn-play"
-            disabled={ this.enableButton() }
-            onClick={ () => {
-              doFormLogin({ name, email });
-              this.getAndSaveToken();
-            } }
-          >
-            Jogar
-          </button>
-        </Link>
+        <button
+          type="button"
+          data-testid="btn-play"
+          disabled={ this.enableButton() }
+          onClick={ () => {
+            doFormLogin({ name, email });
+            this.getAndSaveQuestions();
+          } }
+        >
+          Jogar
+        </button>
         <Link to="/settings" data-testid="btn-settings">
           <button type="button">Configurações do jogo</button>
         </Link>
@@ -100,6 +107,7 @@ class Login extends React.Component {
 
 const mapDispatchToProps = (dispatch) => ({
   doFormLogin: (obj) => dispatch(doLogin(obj)),
+  saveQuestions: (questions) => dispatch(getQuestionsToStore(questions)),
 });
 
 const mapStateToProps = (state) => ({
@@ -109,6 +117,6 @@ const mapStateToProps = (state) => ({
 
 Login.propTypes = {
   doFormLogin: PropTypes.func.isRequired,
+  saveQuestions: PropTypes.func.isRequired,
 };
-
 export default connect(mapStateToProps, mapDispatchToProps)(Login);
